plugins/dance_bot/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wizard101-dancebot"
version = "0.1.0"
description = "Wizard101 Pet Dance automation bot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "opencv-python==4.8.0.76",
  "Pillow==10.0.1",
  "PyAutoGUI==0.9.54",
  "PyGetWindow==0.0.9",
  "pynput==1.7.6",
  "python-xlib==0.33",
  "mss==9.0.1",
  "numpy==1.25.1",
  "pyscreenshot==3.1",
  "PyMsgBox==1.0.9",
  "pyperclip==1.8.2",
  "PyRect==0.2.0",
  "PyScreeze==0.1.29",
  "pytweening==1.0.7",
  "pywin32-ctypes==0.2.2",
]

[project.urls]
Homepage = "https://github.com/kennyhngo/Wizard101_DanceBot"

[project.scripts]
dancebot = "wizard101_dancebot.cli:main"

[tool.setuptools.packages.find]
include = ["wizard101_dancebot*"]

[tool.setuptools.package-data]
"wizard101_dancebot" = ["assets/**/*"]
